Extract Bounce from encfstools to avoid circular import
